Исправления №2
