+++ Content.Shared/Imperial/ImperialBorgs/BorgHypoComponent.cs
@@ -1,15 +1,15 @@
-using Content.Shared.Chemistry.Components;
 using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization;
+using Robust.Shared.Utility;
 
 namespace Content.Shared.Imperial.ImperialBorgs
 {
     [RegisterComponent, NetworkedComponent]
     public sealed partial class BorgHypoComponent : Component
     {
-        [DataField("Solutions")]
-        public List<BorgSolution> Solutions = new List<BorgSolution>();
+        [DataField("solutions")]
+        public List<BorgHypoSolution> Solutions = new();
 
         public int CurrentIndex = 0;
 
@@ -28,45 +28,15 @@ namespace Content.Shared.Imperial.ImperialBorgs
 
     [DataDefinition]
     [Serializable, NetSerializable]
-    public sealed partial class BorgSolution
-    {
-        [DataField("reagents")]
-        public List<ImperialBorgsReagent> Reagents = new();
-
-        public string? GetPrimaryReagentId()
-        {
-            return Reagents.Count > 0 ? Reagents[0].ReagentId : null;
-        }
-
-        public Solution ToChemSolution()
-        {
-            var solution = new Solution();
-            foreach (var reagent in Reagents)
-            {
-                solution.AddReagent(reagent.ReagentId, reagent.Quantity);
-            }
-            return solution;
-        }
-    }
-
-    [DataDefinition]
-    [Serializable, NetSerializable]
     public sealed partial class ImperialBorgsReagent
     {
-        [DataField("ReagentId")]
+        [DataField("reagentId")]
         public string ReagentId = null!;
 
-        [DataField("Quantity")]
+        [DataField("quantity")]
         public float Quantity = 1.0f;
 
-        [DataField("Sprite")]
-        public string? Sprite;
-    }
-
-    [Serializable, NetSerializable]
-    public sealed class BorgHypoComponentState(bool uiUpdateNeeded, string currentReagenName) : ComponentState
-    {
-        public readonly bool UiUpdateNeeded = uiUpdateNeeded;
-        public readonly string CurrentReagentName = currentReagenName;
+        [DataField("sprite")]
+        public ResPath? Sprite;
     }
 }
