+++ Content.Shared/Imperial/ImperialBorgs/Events/BorgHypoComponentState.cs
@@ -1,0 +1,10 @@
+using Robust.Shared.Serialization;
+
+namespace Content.Shared.Imperial.ImperialBorgs.Events;
+
+[Serializable, NetSerializable]
+public sealed class BorgHypoComponentState(bool uiUpdateNeeded, string currentReagenName) : ComponentState
+{
+    public readonly bool UiUpdateNeeded = uiUpdateNeeded;
+    public readonly string CurrentReagentName = currentReagenName;
+}
