+++ Resources/Prototypes/Imperial/Roles/Jobs/Synthetics/ImperialBorgModules.yml
@@ -74,31 +74,31 @@
   - type: StaticPrice
     price: 750
   - type: BorgHypo
-    Solutions:
+    solutions:
       - reagents:
-        - ReagentId: Bicaridine
-          Quantity: 1
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/brutekit.png
+        - reagentId: Bicaridine
+          quantity: 1
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/brutekit.png
       - reagents:
-        - ReagentId: Dermaline
-          Quantity: 1
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/burnkit.png
+        - reagentId: Dermaline
+          quantity: 1
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/burnkit.png
       - reagents:
-        - ReagentId: Dexalin
-          Quantity: 1
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/o2kit.png
+        - reagentId: Dexalin
+          quantity: 1
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/o2kit.png
       - reagents:
-        - ReagentId: Inaprovaline
-          Quantity: 1
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/firstaid.png
+        - reagentId: Inaprovaline
+          quantity: 1
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/firstaid.png
       - reagents:
-        - ReagentId: Hyronalin
-          Quantity: 1
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/radkit.png
+        - reagentId: Hyronalin
+          quantity: 1
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/radkit.png
       - reagents:
-        - ReagentId: Dylovene
-          Quantity: 1
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/toxinkit.png
+        - reagentId: Dylovene
+          quantity: 1
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/toxinkit.png
   - type: SolutionRegeneration
     solution: hypospray
     generated:
@@ -178,31 +178,31 @@
   - type: UseDelay
     delay: 0.5
   - type: BorgHypo
-    Solutions:
+    solutions:
       - reagents:
-        - ReagentId: Bicaridine
-          Quantity: 2
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/brutekit.png
+        - reagentId: Bicaridine
+          quantity: 2
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/brutekit.png
       - reagents:
-        - ReagentId: Dermaline
-          Quantity: 2
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/burnkit.png
+        - reagentId: Dermaline
+          quantity: 2
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/burnkit.png
       - reagents:
-        - ReagentId: Dexalin
-          Quantity: 2
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/o2kit.png
+        - reagentId: Dexalin
+          quantity: 2
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/o2kit.png
       - reagents:
-        - ReagentId: Inaprovaline
-          Quantity: 2
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/firstaid.png
+        - reagentId: Inaprovaline
+          quantity: 2
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/firstaid.png
       - reagents:
-        - ReagentId: Hyronalin
-          Quantity: 2
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/radkit.png
+        - reagentId: Hyronalin
+          quantity: 2
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/radkit.png
       - reagents:
-        - ReagentId: Dylovene
-          Quantity: 2
-          Sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/toxinkit.png
+        - reagentId: Dylovene
+          quantity: 2
+          sprite: /Textures/Objects/Specific/Medical/firstaidkits.rsi/toxinkit.png
   - type: SolutionRegeneration
     solution: hypospray
     generated:
